% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_maps.R
\name{clim_maps}
\alias{clim_maps}
\title{climatic maps}
\usage{
clim_maps(cve_ent, cve_mpo = NULL, state)
}
\arguments{
\item{cve_ent}{is the state id.}

\item{cve_mpo}{is the municipality id.}

\item{state}{is logical value (TRUE or FALSE). If state is TRUE, the map is state leval, else the map is a municipality level.}
}
\value{
a ggplot.
}
\description{
climatic maps
}
\author{
felipe.dzul.m@gmail.com
}
