% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mort_year_edo_dt.R
\name{mort_year_edo_dt}
\alias{mort_year_edo_dt}
\title{Mortalidad by State, year, & age group.}
\usage{
mort_year_edo_dt(dataset, years, cve_edo, cat_group = NULL)
}
\arguments{
\item{dataset}{is the mortality dataset.}

\item{years}{is the year or years analyzed.}

\item{cve_edo}{is the state id.}

\item{cat_group}{is the categorical age group. The options are "productiva", "infantil", "pediatric", "pos-productiva". The function provide the general mortality when cat_group is NULL.}
}
\value{
a kabble object (table).
}
\description{
Mortalidad by State, year, & age group.
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
