% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{mortalidad_general_21}
\alias{mortalidad_general_21}
\title{Mortality data of Puebla state.}
\format{
A dataframe object with 284000 rows and 16 variables:
\describe{
  \item{ent_ocurr}{state id. Two digits, Puebla state is "21".}
  \item{mun_ocurr}{municipality id. Three digits.}
  \item{loc_ocurr}{locality id. Four digits.}
  \item{causa_def_cve}{cause of death identifier.}
  \item{lista_mex_cve}{mexican list id.}
  \item{sexo}{sexo of death.}
  \item{edad_cve}{age id.}
  \item{anio_ocur}{year of occurrence.}
  \item{anio_nacim}{year of birth.}
  \item{ocupacion_cve}{occupation identifier.}
  \item{escolaridad_cve}{scholarship identifier.}
  \item{causa_def_des}{description of cause of death.}
  \item{escolaridad_des}{description of schooling.}
  \item{ocupacion_des}{description of the occupation.}
  \item{edad_des}{description of the age.}
  \item{lista_mex_des}{description of mexican list.}
  ...
}
}
\source{
\url{https://www.inegi.org.mx/programas/mortalidad/#Datos_abiertos}
}
\usage{
mortalidad_general_21
}
\description{
A dataset containing the general mortality data  (2012-2019)
from state of Puebla.
}
\keyword{datasets}
