% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temprec_maps.R
\name{temprec_maps}
\alias{temprec_maps}
\title{Temperature and precipitation maps}
\usage{
temprec_maps(
  cve_mpo,
  cve_ent,
  locality,
  raster_object,
  legend_title,
  width_legend
)
}
\arguments{
\item{cve_mpo}{is the municipality id.}

\item{cve_ent}{is the state id.}

\item{locality}{is the name of locality.}

\item{raster_object}{is the precipitation or temperature raster object.}

\item{legend_title}{is the legend title.}

\item{width_legend}{is width of legend.}
}
\value{
a ggplot object.
}
\description{
Temperature and precipitation maps
}
\author{
felipe.dzul.@gmail.com
}
