% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mort_year_edo_map.R
\name{mort_year_edo_map}
\alias{mort_year_edo_map}
\title{mortality maps by state}
\usage{
mort_year_edo_map(
  dataset,
  years,
  cve_edo,
  cat_group = NULL,
  cve_def = NULL,
  static_map,
  by_mun
)
}
\arguments{
\item{dataset}{is the mortality dataset.}

\item{years}{is the year or years analyzed.}

\item{cve_edo}{is the state id.}

\item{cat_group}{is the categorical age group. The options are "productiva", "infantil", "pediatric", "pos-productiva". The function provide the general mortality when cat_group is NULL.}

\item{cve_def}{is the id of cause of mortality.}

\item{static_map}{is logical valur (TRUE or FALSE), if static_map is TRUE the map is static (ggplot2), else the maps es interecative (leafletmap)}

\item{by_mun}{is a logical value (TRUE or FALSE). If TRUE the level is municipality, else localities.}
}
\value{
a ggplot or leaflet map.
}
\description{
mortality maps by state
}
